Fix u16 overflow panic in call-family dispatch handlers
frame.base + base_reg was computed as u16 + u16 before the bounds check,
causing a checked overflow panic in ReleaseSafe on malformed .sbc files
with large base_reg operands. Widen both operands to usize before adding,
and let ensureCallWindow reject out-of-range values gracefully.

Closes #512

+++ src/vm.zig
@@ -925,7 +925,7 @@ pub const VM = struct {
         return vm_dispatch.registerIndex(self, base, reg);
     }
 
-    fn ensureCallWindow(self: *VM, base: u16, nargs: u8) VMError!void {
+    fn ensureCallWindow(self: *VM, base: usize, nargs: u8) VMError!void {
         return vm_dispatch.ensureCallWindow(self, base, nargs);
     }
